move redis operations in main.go to modules/buildlist.go
