+++ handler/build.go
@@ -1,9 +1,11 @@
 package handler
 
 import (
-	"github.com/Unknwon/macaron"
+//	"github.com/Unknwon/macaron"
 )
 
+/*
 func IndexHandler(ctx *macaron.Context) {
 	ctx.HTML(200, "index")
 }
+*/
