+++ main.go
@@ -1,86 +1,15 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
-	"runtime"
 
 	"github.com/codegangsta/cli"
 	"github.com/containerops/generator/cmd"
+	"github.com/containerops/generator/modules"
 	"github.com/containerops/wrench/setting"
-	"github.com/garyburd/redigo/redis"
 )
-
-/*
-func init() {
-
-	if err := setting.SetConfig("conf/containerops.conf"); err != nil {
-		fmt.Printf("Read config error: %v", err.Error())
-	}
-
-	result := readConfigFile("./conf/pool.json")
-	var list BuilderList
-	if err := json.Unmarshal([]byte(result), &list); err != nil {
-		log.Fatalln(err)
-	} else {
-		saveToRedis(list)
-	}
-
-}
-*/
-
-//move to modules
-func readConfigFile(path string) string {
-	file, err := os.Open(path)
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-	result, err := ioutil.ReadAll(file)
-	if err != nil {
-		panic(err)
-	}
-	return string(result)
-}
-
-//move to modules
-type BuilderList struct {
-	Dockers []BuilderInfo `json:"docker"`
-	Rkt     []BuilderInfo `json:"rkt"`
-}
-
-//move to modules
-type BuilderInfo struct {
-	IP   string `json:"url"`
-	PORT string `json:"port"`
-}
-
-//move to modules
-func saveToRedis(list BuilderList) {
-	c, err := redis.Dial("tcp", setting.DBURI, redis.DialPassword(setting.DBPasswd), redis.DialDatabase(int(setting.DBDB)))
-	if err != nil {
-		panic(err)
-	}
-	defer c.Close()
-
-	for _, v := range list.Dockers {
-		_, err = c.Do("SADD", "DockerList", v.IP+":"+v.PORT)
-		if err != nil {
-			log.Println("err when save docker build list", err)
-		}
-	}
-
-	for _, v := range list.Rkt {
-		_, err = c.Do("SADD", "RKTList", v.IP+":"+v.PORT)
-		if err != nil {
-			log.Println("err when save docker build list", err)
-		}
-
-	}
-}
 
 func main() {
 
@@ -88,13 +17,8 @@ func main() {
 		fmt.Printf("Read config error: %v", err.Error())
 	}
 
-	//to one func
-	result := readConfigFile("./conf/pool.json")
-	var list BuilderList
-	if err := json.Unmarshal([]byte(result), &list); err != nil {
-		log.Fatalln(err)
-	} else {
-		saveToRedis(list)
+	if err := modules.LoadBuildList("./conf/pool.json"); err != nil {
+		log.Fatal(err)
 	}
 
 	app := cli.NewApp()
