+++ router/router.go
@@ -7,7 +7,7 @@ import (
 
 func SetRouters(m *macaron.Macaron) {
 
-	m.Get("/", handler.IndexHandler)
+	//m.Get("/", handler.IndexHandler)
 	//m.Get("/ws", handler.WSServer)
 	m.Get("/wsbuildlog", handler.WSbuildLog)
 	m.Post("/httpbuildlog", handler.HTTPBuildLog)
