+++ static/html/index.html
@@ -3,7 +3,7 @@
  <head> 
   <meta name="generator" content="HTML Tidy for Linux (vers 25 March 2009), see www.w3.org" /> 
   <script src="http://code.jquery.com/jquery-2.1.4.min.js" type="text/javascript"></script>
-  <script src="js/base64.js" type="text/javascript"></script> 
+  <script src="http://192.168.199.10:8080/js/base64.js" type="text/javascript"></script> 
   <title>websokcet test</title> 
  </head> 
  <body> 
@@ -29,7 +29,7 @@ $(function() {
 
     sock.onclose = function(e) {
         console.log("connection closed (" + e.code + ")");
-        appendLog($("<div><b>Connection closed.</b></div>"));
+        // appendLog($("<div><b>Connection closed.</b></div>"));
     }
 
     sock.onmessage = function(e) {
